requirements.txt
fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.23.0,<0.28.0
transformers>=4.38.0,<4.46.0
torch>=2.1.0,<2.3.0
sentence-transformers>=2.3.0,<3.0.0
spacy>=3.7.0,<3.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
faiss-cpu>=1.7.4,<1.8.0
pydantic>=2.6.0,<3.0.0
